requirements.txt
Flask==2.0.1
Flask-Limiter==2.4.0
pdf2docx==0.5.6
pymupdf==1.19.6
pytesseract==0.3.9
Pillow==9.0.1
python-docx==0.8.11
numpy==1.21.2
requests==2.26.0
Werkzeug==2.0.1